% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTourn.R
\name{PlotTourn}
\alias{PlotTourn}
\title{Draws a Blank March Madness Bracket}
\usage{
PlotTourn(games = 32)
}
\arguments{
\item{games:}{number of games to build a bracket for (32 is the default)}

\item{playin:}{not available yet}
}
\value{
a ggplot object of a blank march madness tournament
}
\description{
Draws a blank bracket March Madness bracket as a ggplot object
}
