% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preprocess}
\alias{preprocess}
\title{Method of getting data into a format to be read into}
\usage{
preprocess(year, teamseeds, teamnames)
}
\arguments{
\item{year:}{year of data you want to train on}

\item{teamseeds:}{filepath of teams and their seeds}

\item{teamnames:}{filepath of team id to team name mapping}
}
\value{
data table ready for simulation
}
\description{
Method of getting data into a format to be read into
}
