% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTourn.R
\name{plotTourn}
\alias{plotTourn}
\title{Draws a Blank March Madness Bracket}
\usage{
plotTourn(games = 32, playin = NA)
}
\arguments{
\item{games:}{integer, number of games to build a bracket for (32 is the default)}

\item{playin:}{vector of games that have a play-in game preceding it. playin =
NA, then draws graph with no playin game}
}
\value{
a ggplot object of a blank march madness tournament
}
\description{
Draws a blank bracket March Madness bracket as a ggplot object
}
